Refactor and start daemon via terminal
