+++ intellij-plugin/build.gradle.kts
@@ -27,7 +27,7 @@ intellij {
    version.set("2023.2.8")
    type.set("IC") // Target IDE Platform
 
-   plugins.set(listOf(/* Plugin Dependencies */))
+   plugins.set(listOf("terminal"))
 }
 
 tasks {
