--- intellij-plugin/src/main/kotlin/io/github/ethersync/DaemonOutputNotifier.kt
@@ -1,17 +1,0 @@
-package io.github.ethersync
-
-import com.intellij.util.messages.Topic
-
-interface DaemonOutputNotifier {
-   companion object {
-
-      @JvmField
-      @Topic.ProjectLevel
-      val CHANGE_ACTION_TOPIC: Topic<DaemonOutputNotifier> =
-         Topic.create("ethersync daemon output", DaemonOutputNotifier::class.java)
-   }
-
-   fun clear()
-
-   fun logOutput(line: String)
-}
